refactor: split ai chat backend files
